+++ lib/Controller/ApiController.php
@@ -7,6 +7,7 @@ namespace OCA\SynaplanIntegration\Controller;
 use OCA\SynaplanIntegration\AppInfo\Application;
 use OCA\SynaplanIntegration\Service\LanguageService;
 use OCA\SynaplanIntegration\Service\SynaplanClient;
+use OCP\App\IAppManager;
 use OCP\AppFramework\Controller;
 use OCP\AppFramework\Http;
 use OCP\AppFramework\Http\JSONResponse;
@@ -66,6 +67,7 @@ class ApiController extends Controller
         private IUserSession $userSession,
         private LanguageService $languageService,
         private IConfig $config,
+        private IAppManager $appManager,
         private LoggerInterface $logger,
     ) {
         parent::__construct(Application::APP_ID, $request);
@@ -95,6 +97,7 @@ class ApiController extends Controller
 
         return new JSONResponse([
             'success' => true,
+            'version' => $this->appManager->getAppVersion(Application::APP_ID),
             'language' => $language,
             'languageName' => $this->languageService->getLanguageName($language),
             'useInterfaceLanguage' => $this->languageService->useInterfaceLanguage(),